incoming_archives/files-complete/requirements.txt
# ── MCP Server ───────────────────────────────────────────────────────────────
mcp>=1.0.0
python-dotenv>=1.0.0

# ── Jira Client ───────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0

# ── Embeddings ────────────────────────────────────────────────────────────────
sentence-transformers>=3.0.0

# ── Vector Store ─────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── RAG / Crawling ────────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# ── PDF Indexing ──────────────────────────────────────────────────────────────
pypdf>=4.0.0

# ── XML Validation ────────────────────────────────────────────────────────────
lxml>=5.0.0

# ── Web Scraping ──────────────────────────────────────────────────────────────
beautifulsoup4>=4.12.0

# ── API Server (if running as team HTTP server) ───────────────────────────────
fastapi>=0.111.0
uvicorn>=0.30.0

# ── Fine-tuning (optional) ────────────────────────────────────────────────────
# Uncomment if you want to fine-tune embeddings
# datasets>=2.20.0

# ── Playwright (optional, for richer crawling) ────────────────────────────────
# Uncomment and run: playwright install chromium
# playwright>=1.44.0

# ── Postgres driver (optional, if not using SQLite) ──────────────────────────
# Uncomment if using PostgreSQL
# psycopg2-binary>=2.9.0
